server: initial attempt at DIAL support. This adds a new "DIAL mode" to the server, which can be enabled using the "-d" argument.
When enabled, the server will still take initial commands the same way (A,B,C,... to indicate which half disk is being accessed), but the arguments received after use the DIAL device-specific format:
1) unit number (+ read/write flag)
2) Core location divided by 400 (octal)
3) First block number
4) Block count

DIAL does not encode the write flag into the unit number, but as far as I can tell, nothing uses bit 0 of the unit number, so it seemed to be a convenient place to shove the write flag.
The core location also includes the field number, and, unlike os/8, there are fewer restrictions on how many blocks can be read/written.
The PDP-side handler will handle reads/writes that extend beyond the end of a field by looping back to the start of that field.
So accesses larger than 40 (octal) will simply overread/write themselves (this is what the official LINCtape unit handler code does).